Issue #657 - PR comment changes
- moved package mutexes into packagecommon.go
- reworked construction of "request already in progress" error messages
- replaced cachedRepository polling with retrying the PackageRev update
  to allow PackageRev deletion without redundant polling

https://github.com/nephio-project/nephio/issues/657
