+++ pkg/cache/repository.go
@@ -319,7 +319,6 @@ func (r *cachedRepository) DeletePackage(ctx context.Context, old repository.Pac
 }
 
 func (r *cachedRepository) Close() error {
-	r.pollOnce(context.TODO())
 	r.cancel()
 
 	// Make sure that watch events are sent for packagerevisions that are
@@ -339,7 +338,7 @@ func (r *cachedRepository) Close() error {
 			// There isn't much use in returning an error here, so we just log it
 			// and create a PackageRevisionMeta with just name and namespace. This
 			// makes sure that the Delete event is sent.
-			klog.Warningf("Error deleting PackageRev CR for %s: %v", nn.Name, err)
+			klog.Warningf("repo %s: error deleting packagerev for %s: %v", r.id, nn.Name, err)
 			pkgRevMeta = meta.PackageRevisionMeta{
 				Name:      nn.Name,
 				Namespace: nn.Namespace,
